Fix type to float instead of int

+++ 0x02-python_async_comprehension/2-measure_runtime.py
@@ -6,7 +6,7 @@ import time
 async_comprehension = __import__('1-async_comprehension').async_comprehension
 
 
-async def measure_runtime() -> int:
+async def measure_runtime() -> float:
     """Function that measures total runtime of async functions"""
 
     tasks = []
